requirements.txt
requests
pymongo
tqdm
python-dotenv
selenium
cloudscraper